Add files via upload
The higher order dimensions algorithm based on cube replication and extrapolation of different sets. The main idea is to compute in Scikit-python the high dimensions to recreate new mathematical and computational scenarios with python. In this example, I show how algorithmic logarithmic regression works to elucidate some important machine learning concepts for Scikit and to propose another level of computational analysis.