Add files via upload
Qubits in higher dimension states:

In principle, the theory based on quantum dynamics and quantum cryptography ,the qubits
Are placed in vectors in different dimensional spaces.

The main objective is build a comprehensive mathematical and programming model in python
To stay close to the idea to change the paradigm of traditional computers and robots breaking
Kinematics and motion and give a more amplitude of quantum computing and higher dimensions
In cube states because it is necessary to move forward to further visions and steps to formulate in a concise manner with examples how to apply a new computational model to the physical computers. Usually we use computers with 2D-3D or in planar spaces. What I want to achieve
Is build a computational model that breaks the 2D-3D dimensional states to work similar as the space algorithms in a given multidimensional system is actually working 

For example when you have a qubit in a given space z, the resolution states occur simultaneously across different dimensions. Let's unify quantum cryptography models with higher order dimensions and try to compute it with python in further examples.